\name{DR.SC-package}
\alias{DR.SC-package}
\alias{DR.SC}
\docType{package}
\title{
\packageTitle{DR.SC}
}
\description{
\packageDescription{DR.SC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DR.SC}
\packageIndices{DR.SC}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DR.SC}

Maintainer: \packageMaintainer{DR.SC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
