% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAdj_reg}
\alias{getAdj_reg}
\title{Calculate Adjonit matrix for regular-distributed spots}
\usage{
  getAdj_reg(pos, platform ='Visisum')
}
\arguments{
  \item{pos}{is a n-by-2 matrix of position.}
  
  \item{platform}{a string, specify the platform of the provided data, default as "Visium".}
}
\value{
  A sparse matrix containing the neighbourhood
}
\seealso{
  \link{getAdj_auto}, \link{getAdj}, \link{getAdj_manual}.
}
\description{
  an efficient function to find the neighborhood based on the matrix of position for the platform with regular-distributed spots, such as 'Visium' and 'ST'.
}
