\name{dlpfc151510}
\alias{dlpfc151510}
\title{A human dorsolateral prefrontal cortex data}
\description{
  A human dorsolateral prefrontal cortex dataset measured on the  Visium platform, which invcludes 4634 spots and 500 genes, a subset of raw dataset at https://github.com/LieberInstitute/spatialLIBD.
}

\references{
  None
}
\author{
  Wei Liu
}
\note{
  nothing
}
\examples{
  data("dlpfc151510")
}
