\name{seu}
\alias{seu}
\title{A simulated spatial transcriptomics data}
\description{
  This data is a simulated spatial transcriptomics data based on ST platform, has 100 spots and 50 features.
}

\references{
  None
}
\author{
  Wei Liu
}
\note{
  nothing
}
\examples{
  data(seu)
}
